+++ package.json
@@ -19,7 +19,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router": "^6.16.0",
-    "react-router-dom": "^6.16.0"
+    "react-router-dom": "^6.16.0",
+    "vike": "^0.4.158"
   },
   "devDependencies": {
     "@types/react": "^18.2.15",
