+++ src/App.css
@@ -1,25 +1,15 @@
 @import url('https://fonts.googleapis.com/css2?family=Young+Serif&display=swap');
 
-html {
-  font-family: Inter,Open Sans,Arial,sans-serif;
-  overflow: 'hidden';
-  font-weight: 600px;
-  background-repeat: no-repeat;  
-  
-}
-body > #root{
-  height: 100vh;
-  background-image: url('../src/assets/nnnoise.svg');
-  background-repeat: repeat;
-}
 
-#root {
-  background-color: rgb(248, 248, 248);
-  opacity: 0.8;
+body {
+  background-image: "/src/assets/paper-texture.jpg";
   margin-left: auto;
   margin-right: auto;
+  overflow: 'hidden';
+  font-weight: 600px;
 } 
 
 #UserProfile {
   text-align: right;
 }
+
