+++ src/actions/Search.jsx
@@ -1,0 +1,12 @@
+import React from {React};
+
+
+export default function Search(){
+
+    return (
+	<>
+	    
+	</>
+    )
+    
+}
