--- src/components/BackButton.jsx
@@ -1,16 +1,0 @@
-//import { useHistory } from 'react-router-dom';
-
-export default function BackButton () {
-  //const history = useHistory();
-
- //const goBack = () => {
-  //  history.goBack();
- // };
-
- // return (
- //   <button onClick={goBack}>
- //     {history.length > 1 ? 'Go Back' : 'Home'}
- //   </button>
-// );
-}
-
