+++ src/components/LoadingPage.jsx
@@ -1,0 +1,11 @@
+import '../styles/loadingPage.css'
+
+
+export default function LoadingPage() {
+    return (
+        <>
+        <h1>Loading...</h1>
+        <div id="lds-circle"><div></div></div>
+        </>
+    );
+}
