+++ src/components/ViewPost.jsx
@@ -1,15 +1,15 @@
-//import { useState, useEffect } from "react";
-import './styles/ViewPost.css';
-import { useState } from 'react';
+import { useState, useEffect} from "react";
+import '../styles/ViewPost.css';
 import client from '../lib/pocketbase';
-import { useEffect } from 'react';
 //import PostText from './CreatePost';
-//import axios from 'axios';
 import pb from '../lib/pocketbase'
-import bookmark from '../hooks/bookmark';
+
 
 export default function PostFunction() {
   const [posts, setPosts] = useState([])
+
+  // figure out how to add bookmarking latr
+  //const [bookmarked, setBookMarked] = useState(False)
   
     
   //load data from database, and then render using useEffect on page load
@@ -20,15 +20,23 @@ export default function PostFunction() {
       .then((res) => setPosts(res));
   }, [])
 
+    
+    //async function record() { await pb.collection('posts').update('Book_Mark', true); }
+    
+
 
   return (
     <>
+
     
+    
+
       {posts.map((posts) => (
       <div id='PostViewDiv' key={posts.id}>
-        <input id='PostBookMark'  type='checkbox' onChange={bookmark}></input>
         <h3 id='PostTitle'>{posts.Post_Title}</h3>
         <p id= 'PostText'>{posts.Post_Text}</p>
+        {/*<button id='BookMark' onChange={record()}>BookMark</button>*/}
+        
         
         </div>
       ))}
