--- src/components/styles/NavBar.css
@@ -1,26 +1,0 @@
-#Navigator {
-  padding: 8px 16px;
-  font-weight: bolder;
-}
-
-ul {
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-  width: 25%;
-  position: fixed;
-  height: 100%;
-  overflow: auto;
-
-}
-
-li a {
-  display: block;
-  color: #000;
-  padding: 8px 16px;
-  text-decoration: none;
-}
-
-li a:hover:not(.active) {
-  background-color: #f1f1f1;
-  color: rgb(0, 0, 0);
