--- src/components/styles/Post.css
@@ -1,93 +1,0 @@
-
-@import url('https://fonts.googleapis.com/css2?family=Young+Serif&display=swap');
-
-#post {
-    max-width: 1280px;
-    font-family: Inter,Open Sans,Arial,sans-serif;
-    padding: 2rem;
-    text-align: center;
-}  
-
-#titleAreaText {
-  position: relative;
-  overflow-x: auto; 
-    overflow-y: hidden;
-  padding-bottom: 5px;
-  padding-top: 5px;
-  margin-bottom: 2%;
-    font-family: Inter,Open Sans,Arial,sans-serif;
-    resize: none;
-    width: 200px;
-    height: 32px;
-    font-size: 16px;
-    border-radius: 10px;
-    background-color: rgba(255, 255, 255, 0);
-    padding: 10px;
-    margin-left: auto;
-    margin-right: auto;
-    display: flex;
-    justify-content: left;
-    flex-wrap: nowrap;
-    flex-direction: row;
-}
-
-  #PostTextArea {
-    margin-bottom: 5%;
-    font-family: Inter,Open Sans,Arial,sans-serif;
-    resize: none;
-    width: 300px;
-    height: 170px;
-    font-size: 14px;
-    border-radius: 20px;
-    background-color: rgba(255, 255, 255, 0);
-    padding: 20px;
-    margin-left: auto;
-    margin-right: auto;
-    display: flex;
-    justify-content: center;
-    flex-wrap: nowrap;
-    flex-direction: row;
-  }
-  
-  #textPost {
-    margin-bottom: 5%;
-    border-radius: 20px;
-    margin: 0, auto;
-    padding: 5rem;
-    text-align: center;
-    max-width: fit-content;
-    height: fit-content;
-    background-color: #58F3Af;
-    color: black;
-  }
-
-#upload {
-  stroke: black;
-  stroke-width: 1px;
-  margin-bottom: 3%;
-  font-family: Inter,Open Sans,Arial,sans-serif;
-  background-color: rgba(255, 255, 255, 0.5);
-  width: 50px;
-  
-  max-width: 100x;
-  text-align: center;
-  margin-left: auto;
-  margin-right: auto;
-  display: flex;
-  justify-content: center;
-  flex-wrap: nowrap;
-  flex-direction: row;
-  border-radius: 20px;
-  border: 1px solid transparent;
-  padding: 0.7% 1%;
-  transition: border-color 0.25s;
-}
-
-
-button:hover {
-  border-color: #646cff;
-}
-button:focus,
-button:focus-visible {
-  outline: 4px auto -webkit-focus-ring-color;
-}
