--- src/components/styles/ViewPost.css
@@ -1,36 +1,0 @@
-#root {
-    color: black;
-    
-}
-
-#PostViewDiv {
-    overflow: hidden;
-    background-color: rgba(255, 255, 255, 0.309);
-    margin-left: auto;
-    margin-right: auto;
-    margin-bottom: 3%;
-    border-radius: 20px;
-    margin: 0, auto;
-    padding: 1rem;
-    min-width: 200px;
-    max-width: fit-content;
-    min-height: 100px;
-    height: fit-content;
-    stroke: black;
-    color: black;
-    
-}
-
-#PostTitle {
-    text-align: left;
-    padding-left: 2%;
-}
-
-#PostText {
-    text-align: left;
-    padding-left: 2%;
-}
-
-#PostBookMark {
-    float: right;
-}
