+++ src/styles/ViewPost.css
@@ -1,0 +1,42 @@
+#root {
+    color: black;
+    
+}
+
+#PostViewDiv {
+    overflow: hidden;
+    background-color: #F4F2EF;
+    margin-left: auto;
+    margin-right: auto;
+    margin-bottom: 3%;
+    border-radius: 20px;
+    margin: 0, auto;
+    padding: 1rem;
+    min-width: 200px;
+    max-width: fit-content;
+    min-height: 100px;
+    height: fit-content;
+    stroke: black;
+    color: black;
+
+    transition-property: background-color;
+    transition-duration: 0.89s;
+    transition-delay: background-color .4 ease
+    
+}
+
+#PostViewDiv:hover {
+    background-color: #F44027;
+    color: #F4F2EF;
+}
+
+#PostTitle {
+    text-align: left;
+    padding-left: 2%;
+}
+
+#PostText {
+    text-align: left;
+    padding-left: 2%;
+}
+
