+++ src/styles/loadingPage.css
@@ -1,0 +1,36 @@
+#lds-circle {
+    display: inline-block;
+    transform: translateZ(1px);
+  }
+  #lds-circle > div {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    position:absolute;
+    margin: auto;
+    border: 3px solid rgb(0, 0, 0);
+    padding: 10px;
+    display: inline-block;
+    width: 64px;
+    height: 64px;
+    margin: 8px;
+    border-radius: 50%;
+    background: #fff;
+    animation: lds-circle 2.4s cubic-bezier(0, 0.2, 0.8, 1) infinite;
+  }
+  @keyframes lds-circle {
+    0%, 100% {
+      animation-timing-function: cubic-bezier(0.5, 0, 1, 0.5);
+    }
+    0% {
+      transform: rotateY(0deg);
+    }
+    50% {
+      transform: rotateY(1000deg);
+      animation-timing-function: cubic-bezier(0, 0.5, 0.5, 1);
+    }
+    100% {
+      transform: rotateY(1500deg);
+    }
+  }
+
