+++ vite.config.js
@@ -4,4 +4,8 @@ import react from '@vitejs/plugin-react-swc'
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+
+  optimizeDeps: {
+    exclude: ['js-big-decimal']
+  }
 })
